Şekil Güncellemesi
Dünya'nın şekli tekrardan düzeltildi. Artık her köşe ve kenar tek bir nesne her yerde.

+++ kod/başaşağıderebeyi/soyutkuruluş/ulus/Ulus.java
@@ -13,14 +13,14 @@ import java.util.List;
 
 public class Ulus {
 	public final Dünya dünya;
-	public final int[] envanter;
+	public final float[] envanter;
 	public final List<Şehir> şehirler;
 	public final List<Yol> yollar;
 	public final Color renk;
 
 	public Ulus(final Dünya dünya, final Color renk) {
 		this.dünya = dünya;
-		envanter = new int[Kaynak.DEĞERLER.length];
+		envanter = new float[Kaynak.DEĞERLER.length];
 		şehirler = new ArrayList<>();
 		yollar = new ArrayList<>();
 		this.renk = renk;
